+++ src/main/java/com/diablo3CharViewer/CharacterViewerManager.java
@@ -1,19 +1,17 @@
 package com.diablo3CharViewer;
 
+import com.diablo3CharViewer.api_handlers.AccountHandlerApi;
+import com.diablo3CharViewer.token.FetchToken;
+
 import java.util.Scanner;
 
 public class CharacterViewerManager {
 
-    public void showProfile(GetApiAccount getApiAccount, FetchToken fetchToken) {
+    public void showProfile(Scanner scanner, AccountHandlerApi accountHandlerApi, FetchToken fetchToken) {
 
         System.out.println("Witaj wedrowcze! Podaj battleTag aby wyszukac profil bohatera: ");
-        Scanner scanner = new Scanner(System.in);
         String battleTag = scanner.nextLine();
-        scanner.reset();
-        getApiAccount.setAccount(battleTag);
-        getApiAccount.setRequestURL();
-
-        System.out.println("Zostan na chwile i poczytaj:\n" + getApiAccount.generateRequest(fetchToken) + '\n');
+        System.out.println("Zostan na chwile i poczytaj:\n" + accountHandlerApi.generateRequest(battleTag, fetchToken) + '\n');
     }
 
 }
