+++ src/main/java/com/diablo3CharViewer/api_handlers/AccountHandlerApi.java
@@ -1,0 +1,14 @@
+package com.diablo3CharViewer.api_handlers;
+
+import com.diablo3CharViewer.token.FetchToken;
+import com.diablo3CharViewer.token.Token;
+
+public class AccountHandlerApi {
+
+    //zrobic jako static?
+    //skoro zostala 1 metoda to czy robic z tego klase czy wrzucic ta metode do managera??
+    public String generateRequest(String battleTag, FetchToken fetchToken) { //token przypisywać ze zmiennej. Pobieram token kiedy program startuje i wrzucam go do stałej.
+        return fetchToken.initializeApiServiceGET("https://eu.api.blizzard.com/d3/profile/"
+        + battleTag + "/?locale=pl_PL&access_token=" + Token.getAccess_token());
+    }
+}
